Be more resilient when dealing with Hue lights.
This is a partial fix for issue #10 - where the hue node server
crashes if the hue hub has a bulb that does not report XY color,
such as a GE Link bulb (dimmable only).  This fix will result in
the values (0,0) being reported for the XY color, but the on/off
status is correctly reported and the node server no longer crashes.

(TODO: Create a new node type for non-color dimmable bulbs.)